fix(auth_core): 修改返回参数类型为long

+++ src/main/java/cn/rainier/nian/service/impl/MenuServiceImpl.java
@@ -132,7 +132,7 @@ public class MenuServiceImpl implements MenuService {
 	}
 	
 	@Override
-	public int findChildMenuCount(Integer menuId) {
+	public long findChildMenuCount(Integer menuId) {
 		return menuDao.findChildMenuCount(menuId);
 	}
